feat: enhance parcel creation with destination address support and update related models

+++ ms-shipping/infrastructure/driven-adapters/jpa-repository/src/main/java/co/fastshipping/jpa/repository/ParcelJpaRepository.java
@@ -3,5 +3,5 @@ package co.fastshipping.jpa.repository;
 import co.fastshipping.jpa.entity.ParcelJpaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParcelJpaRepository extends JpaRepository<ParcelJpaEntity,Integer> {
+public interface ParcelJpaRepository extends JpaRepository<ParcelJpaEntity, Long> {
 }
